Autonomous
Chadreversed.xml is a config file. It works with the current bot, and if you put it under Teamcod\res\xml it will show up as a config on the RC app. I suggest using it so that we can easily switch phones if needed.

Both crater and normal path use SeanLinearOpmode rather than LinearOpmode, in order to stop without errors at any time. The only change to linear opmode is that stop() isn't final (it's overwritten in Crater and V3).
AutoIntegrateV3 is the program we've been working on to follow (as close as possible) Safe Auto Paths for the depot side. The main changes are that it works with the current bot (everything backwards) and the Align program I've been working on. It seems to work, but 
AutoIntegrateCrater is basically V3 but with orders and directions switched around to start from the crater.
I also haven't tested any of this on an official field.